chore: initialize store in context
